+++ Riot/Modules/Authentication/AuthenticationViewController.h
@@ -41,5 +41,11 @@
 @property (weak, nonatomic) IBOutlet UIView *homeServerSeparator;
 @property (weak, nonatomic) IBOutlet UIView *identityServerSeparator;
 
+@property (weak, nonatomic) IBOutlet UIView *softLogoutClearDataContainer;
+@property (weak, nonatomic) IBOutlet UILabel *softLogoutClearDataLabel;
+@property (weak, nonatomic) IBOutlet UIButton *softLogoutClearDataButton;
+
+- (void)showCustomHomeserver:(NSString*)homeserver andIdentityServer:(NSString*)identityServer;
+
 @end