+++ Riot/Modules/Room/DataSources/RoomDataSource.m
@@ -196,6 +196,8 @@
     if ([cell isKindOfClass:MXKRoomBubbleTableViewCell.class])
     {
         MXKRoomBubbleTableViewCell *bubbleCell = (MXKRoomBubbleTableViewCell*)cell;
+        [self resetAccessibilityForCell:bubbleCell];
+
         RoomBubbleCellData *cellData = (RoomBubbleCellData*)bubbleCell.bubbleData;
         NSArray *bubbleComponents = cellData.bubbleComponents;
 
@@ -238,12 +240,12 @@
                         continue;
                     }
                 
-                    MXAggregatedReactions* reactions = cellData.reactions[componentEventId];
+                    MXAggregatedReactions* reactions = cellData.reactions[componentEventId].aggregatedReactionsWithNonZeroCount;
                     
                     BubbleReactionsView *reactionsView;
                     
-                    if (reactions && !isCollapsableCellCollapsed)
-                    {
+                    if (!component.event.isRedactedEvent && reactions && !isCollapsableCellCollapsed)
+                    {                        
                         BOOL showAllReactions = [cellData showAllReactionsForEvent:componentEventId];
                         BubbleReactionsViewModel *bubbleReactionsViewModel = [[BubbleReactionsViewModel alloc] initWithAggregatedReactions:reactions
                                                                                                                                    eventId:componentEventId
@@ -507,6 +509,8 @@
         
         // Auto animate the sticker in case of animated gif
         bubbleCell.isAutoAnimatedGif = (cellData.attachment && cellData.attachment.type == MXKAttachmentTypeSticker);
+
+        [self setupAccessibilityForCell:bubbleCell withCellData:cellData];
     }
 
     return cell;
@@ -563,6 +567,35 @@
     [self sendVideo:videoLocalURL withThumbnail:videoThumbnail success:success failure:failure];
 }
 
+
+#pragma - Accessibility
+
+- (void)setupAccessibilityForCell:(MXKRoomBubbleTableViewCell *)cell withCellData:(RoomBubbleCellData*)cellData
+{
+    // Set accessibility only on media. Let VoiceOver automatically manages text messages
+    if (cellData.attachment)
+    {
+        NSString *accessibilityLabel = [cellData accessibilityLabel];
+        if (cell.messageTextView.text.length)
+        {
+            // Files are presented as text with link
+            cell.messageTextView.accessibilityLabel = accessibilityLabel;
+            cell.messageTextView.isAccessibilityElement = YES;
+        }
+        else
+        {
+            cell.attachmentView.accessibilityLabel = accessibilityLabel;
+            cell.attachmentView.isAccessibilityElement = YES;
+        }
+    }
+}
+
+- (void)resetAccessibilityForCell:(MXKRoomBubbleTableViewCell *)cell
+{
+    cell.messageTextView.accessibilityLabel = nil;
+    cell.attachmentView.accessibilityLabel = nil;
+}
+
 #pragma mark - BubbleReactionsViewModelDelegate
 
 - (void)bubbleReactionsViewModel:(BubbleReactionsViewModel *)viewModel didAddReaction:(MXReactionCount *)reactionCount forEventId:(NSString *)eventId
@@ -605,6 +638,11 @@
 
         [self.delegate dataSource:self didCellChange:nil];
     }
+}
+
+- (void)bubbleReactionsViewModel:(BubbleReactionsViewModel *)viewModel didLongPressForEventId:(NSString *)eventId
+{
+    [self.delegate dataSource:self didRecognizeAction:kMXKRoomBubbleCellLongPressOnReactionView inCell:nil userInfo:@{ kMXKRoomBubbleCellEventIdKey: eventId }];
 }
 
 @end